requirements.txt
absl-py==1.3.0
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
audioread==3.0.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
cryptography==38.0.4
cssselect==1.2.0
cycler==0.11.0
databases==0.6.2
debugpy==1.6.3
decorator==5.1.1
Dijkstar==2.6.0
dill==0.3.6
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
executing==1.0.0
fake-useragent==0.1.14
fastapi==0.87.0
fastapi-utils==0.2.1
filelock==3.8.0
flatbuffers==22.10.26
fonttools==4.29.1
future==0.18.2
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
grpcio==1.50.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.15.2
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jose==1.0.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jwt==1.3.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
MechanicalSoup==1.2.0
multitasking==0.0.11
nest-asyncio==1.5.5
networkx==2.7.1
nltk==3.7
numba==0.56.4
numpy==1.22.2
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.3
packaging==21.3
pandas==1.4.4
parse==1.19.0
parso==0.8.3
passlib==1.7.4
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pooch==1.6.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pyee==8.2.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.7
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
python-speech-features==0.6
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
regex==2022.10.31
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@46a603053cfbc0ce3c54d43d7e1ac2427fa82b4d
soundfile==0.11.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.5.0
starlette==0.21.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.18.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.11.0
termcolor==2.1.0
tf2crf==0.1.13
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.25.1
tweepy @ git+https://github.com/tweepy/tweepy.git@3d3d58e1e8627788fe88684a515ef888038f3e3c
twitter-scraper==0.4.4
typeguard==2.13.3
typing_extensions==4.4.0
ujson==5.6.0
urduhack==1.1.1
urllib3==1.26.12
uvicorn==0.20.0
uvloop==0.17.0
w3lib==2.0.1
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
yfinance==0.1.87
zipp==3.10.0
